third_party/WebKit/LayoutTests/fast/text/font-ligature-letter-spacing-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 1 duplicate test name: "Ligature expected not to be applied due to letter spacing."
FAIL Ligature expected not to be applied due to letter spacing. assert_equals: Ligature not applied due to letter spacing. expected 282.96875 but got 138.359375
FAIL Ligature expected not to be applied due to letter spacing. assert_equals: Ligature not applied due to letter spacing. expected 282.96875 but got 138.359375
PASS Non-ligature font feature expected to be applied despite letter spacing.
Harness: the test ran to completion.
